polymarket-arbitrage/requirements.txt
web3>=6.0
aiohttp>=3.8
pandas>=2.0
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.28
